Reorganize API to be easier to add new versions - Part II.
