+++ config/initializers/swagger.rb
@@ -3,5 +3,6 @@ GrapeSwaggerRails.options.app_name = 'Claim for crown court defence API'
 GrapeSwaggerRails.options.doc_expansion = 'list'
 
 GrapeSwaggerRails.options.before_filter do
-  GrapeSwaggerRails.options.url = request.params.key?(:v2) ? '/api/v2/swagger_doc' : '/api/v1/external_users/swagger_doc'
+  version = request.params.fetch(:v, '1').to_i
+  GrapeSwaggerRails.options.url = "/api/v#{version}/swagger_doc"
 end
